Changed the onClick behavior of the first button on the menu to spawn a fighter jet for the player. The fighter jet will be destroyed once it goes out of the screen. The fighter jet prefab is located in /Resources/Prefabs/FighterPlayer
Changed folder structure and moved Mothership.cs into Resources/Scripts
