+++ Assets/Resources/Prefabs.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 50f397f2d12e212488b6f9575918d813
+folderAsset: yes
+timeCreated: 1454764414
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
