+++ Assets/Resources/Prefabs/PlayerFighter.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1f79b9770a554c9408402ec18c95a6b9
+timeCreated: 1454764431
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
